requirements.txt
# requirements.txt
selenium
flask
pymongo
python-dotenv
requests